+++ src/js/src/components/AppHeader.vue
@@ -84,6 +84,9 @@ are only available to Global Administrator or Global Viewer.
     </template>
 
     <template #end>
+      <b-navbar-item v-if="auth" tag="div">
+        <refresh-status></refresh-status>
+      </b-navbar-item>
       <b-navbar-item v-if="proxyAuth" class="navbar-item" @click="logout"
         >Reauthorize
       </b-navbar-item>
@@ -98,8 +101,10 @@ are only available to Global Administrator or Global Viewer.
   import { usePhenixStore } from '@/store.js';
   import { roleAllowed } from '@/utils/rbac.js';
   import axiosInstance from '@/utils/axios.js';
+  import RefreshStatus from '@/components/RefreshStatus.vue';
 
   export default {
+    components: { RefreshStatus },
     setup() {
       return { roleAllowed };
     },